requirements.txt
apify-client
google-genai
pydantic
schedule
python-dotenv
pandas
streamlit
requests
fpdf2
